frontend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "SensorThings map + dashboard + rasters_vectors"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.1",
    "folium>=0.20.0",
    "requests>=2.31.0",
    "shapely>=2.0.0",
    "pyproj>=3.5.0",
    "gunicorn>=21.2",
    "psycopg2-binary",
    "python-dateutil>=2.9.0.post0",

    # added for spatial rasters/vectors
    "python-dotenv>=1.2.1",
    "numpy>=1.26.0",
    "Pillow>=12.0.0",
    "rasterio>=1.4.2",
]
